+++ config/settings/queue.php
@@ -4,5 +4,5 @@ declare(strict_types=1);
 
 return [
     'defaultQueue' => 'default',
-    'sleep' => 1,
+    'timeout' => 5,
 ];