+++ src/Queue/QueueWorkerOptions.php
@@ -8,7 +8,7 @@ final readonly class QueueWorkerOptions
 {
     public function __construct(
         public string $queueName,
-        public int $sleep,
+        public int $timeout,
         public bool $once,
         public int $maxJobs,
         public int $maxTime,